example/requirements.txt
# Your app requirements.
-r ../requirements.txt

# Your app in editable mode.
-e ../

future==0.16.0

Django==1.11.7

# celery
celery[redis]==3.1.25
threadpool==1.3.2  # celery
django-celery==3.2.1


